requirements.txt
bs4
python-telegram-bot~=13.15
requests